+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -6,13 +6,13 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import woowacourse.shoppingcart.dao.CartItemDao;
-import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.exception.InvalidProductException;
 import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.CartItemDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.CustomerDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.ProductDao;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
