+++ src/main/java/woowacourse/shoppingcart/application/OrderService.java
@@ -7,16 +7,16 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import woowacourse.shoppingcart.dao.CartItemDao;
-import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dao.OrderDao;
-import woowacourse.shoppingcart.dao.OrdersDetailDao;
-import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.OrderRequest;
 import woowacourse.shoppingcart.exception.InvalidOrderException;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.CartItemDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.CustomerDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.OrderDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.OrdersDetailDao;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.ProductDao;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
