+++ src/main/java/woowacourse/shoppingcart/application/ProductService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.infrastructure.jdbc.dao.ProductDao;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
